python/requirements.txt
ai-waifu==0.1.0 (from file:///Users/hsnmkls/work/ai-waifu)
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
bitarray==2.9.2
blessed==1.20.0
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.3
cytoolz==0.12.3
dataclasses-json==0.6.4
deprecated==1.2.14
distro==1.9.0
ecdsa==0.18.0
editor==1.6.6
eth-abi==5.0.0
eth-account==0.11.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==3.0.0
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.28.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inquirer==3.2.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langsmith==0.0.87
lru-dict==1.2.0
marshmallow==3.20.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.1.6
onnxruntime==1.17.0
openai==1.12.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.7.0
packaging==23.2
parsimonious==0.9.0
posthog==3.4.1
protobuf==4.25.3
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic-core==2.16.2
pypika==0.48.9
pyproject-hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
pyunormalize==15.1.0
pyyaml==6.0.1
readchar==4.0.5
referencing==0.33.0
regex==2023.12.25
replicate==0.24.0
requests==2.31.0
requests-oauthlib==1.3.1
rlp==4.0.0
rpds-py==0.18.0
rsa==4.9
runs==1.2.2
safetensors==0.4.2
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.27
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.2
transformers==4.37.2
typer==0.9.0
typing-extensions==4.9.0
typing-inspect==0.9.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
web3==6.15.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
xmod==1.8.1
yarl==1.9.4
zipp==3.17.0
